bpf: Avoid implicit shorten-64-to-32 in clang 19
As per the below PR, clang 19.1.x is having -Wshorten-64-to32 under the
-Wimplicit-int-conversion, which is already forbidden in Makefile.bpf.
This commit is to avoid the conversion error as per below sample

```
./lib/nodeport.h:1825:18: error: implicit conversion loses integer precision: '__u64' (aka 'unsigned long long') to '__s32' (aka 'int') [-Werror,-Wshorten-64-to-32]
 1825 |         __s32 len_old = ctx_full_len(ctx);
```

Relates: https://github.com/llvm/llvm-project/pull/80814

+++ bpf/lib/common.h
@@ -198,7 +198,7 @@ ____revalidate_data_pull(struct __ctx_buff *ctx, void **data_, void **data_end_,
 
 	/* Verifier workaround, do this unconditionally: invalid size of register spill. */
 	if (pull)
-		ctx_pull_data(ctx, tot_len);
+		ctx_pull_data(ctx, (__u32)tot_len);
 	data_end = ctx_data_end(ctx);
 	data = ctx_data(ctx);
 	if (data + tot_len > data_end)
@@ -1252,7 +1252,7 @@ static __always_inline int redirect_ep(struct __ctx_buff *ctx __maybe_unused,
 	 */
 	if (needs_backlog || !is_defined(ENABLE_HOST_ROUTING) ||
 	    ctx_get_ingress_ifindex(ctx) == 0) {
-		return ctx_redirect(ctx, ifindex, 0);
+		return (int)ctx_redirect(ctx, ifindex, 0);
 	}
 
 	/* When coming from overlay, we need to set packet type
